+++ ModelSwapperSkins/Main.cs
@@ -13,7 +13,7 @@ namespace ModelSwapperSkins
         public const string PluginGUID = PluginAuthor + "." + PluginName;
         public const string PluginAuthor = "Gorakh";
         public const string PluginName = "ModelSwapperSkins";
-        public const string PluginVersion = "1.5.0";
+        public const string PluginVersion = "1.5.1";
 
         void Awake()
         {